+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Latest Changes
 
+- fix: add event time to SacEvent
 - feat: add Event type and MiniEvent class
 - refactor: move functions into a single file
 - docs: switch to google type docstrings